dtss speed (#222)
* minor fixes and improvements to cache-stats, flush, and some faster handling if all time-series happen to be in cache

* added explicit cache use control for dtss read methods

* adjust defaults (read from cache is default), expose cache_flush to client

* in64_t in serialization interfaces

* remove binary header

* refactor and ensure all serialized ints are int64_t

* ensure core archive flags (noheader) is used everywhere

* cellbuilder with .v3 format

* added new files to ms c++ project

* added missing arch_core_flags

* fixed cb test

* upgraded to boost 1.66 dlib 19.7

* added client-side .cache_stats

* added auto_connect, as default, so socket are opened just before calling, and closed after

* first version of dtss client with multiple server connections

* first working version of multinode

* adjusting tests to study serialization speed

* experimental work in progress using alternate form for serialized expressions

* boost 1.66, work in progress, 2.6x speed for serialization of expressions

* consistent 3x speedup for specialized expr handling, seems worth it

* starting tuple-based structures for expression serialization

* doc tests, and cleanups

* vector(ts) to expr working, reverse direction needs some more work to compile

* complete working test-code based on tuple/variants as representation and engine for flatten/de-flatten expressions as suggested by John Eivind

* added abs_ts

* added expr conversion for average_ts

* added more implementation for expr ts serialization

* ms c++ fix

* convolve added

* added extend_ts

* added rating_curve_ts

* compress code with macro

* krls added

* added qac_ts (that have optional possible empty ts, that brought in boost::blank into the variant type set)

* switching api/time_series.h namespace to shyft::time_series::dd, dynamic_dispatch

* finished moving api/time_series.h to core time_series_dd, including serialization support

* updated codeblocks

* updated cmake build support, now using fixed set of files for shyftcore shyftapi and test project (for speed)

* try older travis image

* hmm. now travis did work without old image, reverting

* added tests to verify  alt. expr. serialization

* added compat flag for boost variant structure to ensure compact layout

* rename and refactor before promoting code out of the test-bench

* moved expression compressor into core

* updated cmake

* removed obsolete empty files and pch

* updated project codeblocks, added kdev

* updated codeblocks projects

* enabled compress_expressions, defaults to true, can be turned off on client side

+++ api/api_state.h
@@ -4,7 +4,7 @@
 #include <vector>
 #include <memory>
 #include <stdexcept>
-#include "core/core_pch.h"
+#include "core/core_serialization.h"
 
 #include "core/geo_cell_data.h"
 #include "core/cell_model.h"
@@ -29,12 +29,12 @@ namespace shyft {
         *
         */
         struct cell_state_id {
-            int cid;///< the catchment id, if entirely different model, this might change
-            int x;///< the cell mid-point x (west-east), truncated to integer [meter]
-            int y;///< the cell mid-point y (south-north), truncated to integer [meter]
-            int area;///< the area in m[m2], - if different cell geometries, this changes
+            int64_t cid;///< the catchment id, if entirely different model, this might change
+            int64_t x;///< the cell mid-point x (west-east), truncated to integer [meter]
+            int64_t y;///< the cell mid-point y (south-north), truncated to integer [meter]
+            int64_t area;///< the area in m[m2], - if different cell geometries, this changes
             cell_state_id() = default; // python exposure
-            cell_state_id(int cid, int x, int y, int area) :cid(cid), x(x), y(y), area(area) {}
+            cell_state_id(int64_t cid, int64_t x, int64_t y, int64_t area) :cid(cid), x(x), y(y), area(area) {}
             bool operator==(const cell_state_id & o) const {
                 return cid == o.cid && x == o.x && y == o.y && area == o.area;
             }
